create channel file
