+++ examples/jsonplaceholder_script.graphql
@@ -1,12 +1,5 @@
 schema
-  @server(
-    port: 8000
-    graphiql: true
-    hostname: "0.0.0.0"
-    script: {
-      path: "examples/scripts/test.js"
-    }
-  )
+  @server(port: 8000, graphiql: true, hostname: "0.0.0.0", script: {path: "examples/scripts/test.js"})
   @upstream(baseURL: "http://jsonplaceholder.typicode.com", httpCache: true) {
   query: Query
 }
