+++ src/app_context.rs
@@ -5,12 +5,13 @@ use async_graphql::Response;
 
 use crate::blueprint::Type::ListType;
 use crate::blueprint::{Blueprint, Definition};
+use crate::channel::{Command, Event};
 use crate::data_loader::DataLoader;
 use crate::graphql::GraphqlDataLoader;
 use crate::grpc::data_loader::GrpcDataLoader;
 use crate::http::{DataLoaderRequest, HttpDataLoader};
 use crate::lambda::{DataLoaderId, Expression, IO};
-use crate::{grpc, Command, EntityCache, EnvIO, Event, HttpIO, ScriptIO};
+use crate::{grpc, EntityCache, EnvIO, HttpIO, ScriptIO};
 
 pub struct AppContext {
   pub schema: dynamic::Schema,
